ibv-conduit: SRQ init was broken due to type changes

+++ ibv-conduit/gasnet_core.c
@@ -977,7 +977,7 @@ static int gasnetc_load_settings(void) {
      GASNET_USE_SRQ > 0: Use SRQ for nodes >= GASNET_USE_SRQ (so 1 == Always)
      GASNET_USE_SRQ < 0: Use SRQ is memory savings would result (done in gasnetc_sndrcv_init()).
    */
-  if (gasnetc_use_srq > gasneti_nodes) {
+  if (gasnetc_use_srq > (int)gasneti_nodes) {
     /* Positive value of GASNET_USE_SRQ denotes a minumum node count */
     GASNETI_TRACE_PRINTF(I, ("SRQ disabled because GASNET_USE_SRQ = %d is greater than nodes = %d",
                              gasnetc_use_srq, gasneti_nodes));
